Microservices/account: Moved files to prepare for major changes
