+++ server/microservices/account/controllers/userController.js
@@ -1,7 +1,7 @@
 import bcrypt from "bcrypt";
 import jwt from "jsonwebtoken";
-import { asyncHandler } from "../../utils/asyncHandler.js";
-import { ErrorResponse } from "../../utils/ErrorResponse.js";
+import { asyncHandler } from "../utils/asyncHandler.js";
+import { ErrorResponse } from "../utils/ErrorResponse.js";
 import { userDBPool } from "../db/index.js";
 
 export const getUser = asyncHandler(async (req, res, next) => {
